requirements.txt
pillow
numpy
pycryptodome
Flask        3.1.0
Flask-Cors   5.0.0